% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{uos_colours}
\alias{uos_colours}
\title{University of Southampton colour palettes}
\format{
An object of class \code{character} of length 19.
}
\usage{
uos_colours
}
\description{
A collection of colour palettes using the University of Southampton colour
scheme. The list of available palettes is:
neutral
marine
horizon
contrast1
contrast2
contrast3
}
\details{
Make a named list of University of Southampton brand colours
}
\examples{
#' # Make an x-y plot using the marine palette
library(tidyverse)
df <- data.frame(x = rnorm(100, 0, 20),
          y = rnorm(100, 0, 20),
          cl = sample(letters[1:8], 100, replace=TRUE))
ggplot(df, aes(x, y, colour=cl, shape=cl)) +
  geom_point(size=4) + scale_colour_uos() +
  theme_bw() + theme(aspect.ratio=1)
}
\keyword{datasets}
