% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{uos_cols}
\alias{uos_cols}
\title{Function to extract uos_colours as hex codes}
\usage{
uos_cols(...)
}
\arguments{
\item{...}{Character names of uos_colours}
}
\description{
Function to extract uos_colours as hex codes
}
