% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{uos_pal}
\alias{uos_pal}
\title{uos palette with ramped colours}
\usage{
uos_pal(palette = "marine", alpha = 1, reverse = FALSE)
}
\arguments{
\item{palette}{Choose from 'uos_palettes' list}

\item{alpha}{transparency}

\item{reverse}{If TRUE, the direction of the colours is reversed.}
}
\description{
uos palette with ramped colours
}
\examples{
\dontrun{
library(scales)
show_col(uos_pal()(10))

filled.contour(volcano,color.palette = uos_pal(), asp=1)
}
#' @export
}
