% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{uos_palettes}
\alias{uos_palettes}
\title{Construct some palettes}
\format{
An object of class \code{list} of length 7.
}
\usage{
uos_palettes
}
\description{
Construct some palettes
}
\keyword{datasets}
